- change names of friction-less pendulum training scripts - change some hyperparameters - fix best weights saving - remove free fall example
