+++ examples/pendulum_friction-less/model_train.jl
@@ -1,3 +1,6 @@
+# Example of GOKU-net model on friction-less pendulum data created with Luxor
+
+using Flux: length
 using LatentDiffEq
 using FileIO
 using Parameters: @with_kw
@@ -9,42 +12,43 @@ using BSON: @save
 using Flux.Data: DataLoader
 using Flux
 using OrdinaryDiffEq
+using StochasticDiffEq
 using ModelingToolkit
 using Images
 using Plots
 import GR
+
+include("create_data.jl")
 
 ################################################################################
 ## Arguments for the train function
 @with_kw mutable struct Args
     ## Global model
     model_type = GOKU()
-    # model_type = LatentODE()
 
     ## Latent Differential Equations
     diffeq = Pendulum()
-    # diffeq = NODE(2)
 
     ## Training params
-    η = 1e-3                        # learning rate
+    η = 5e-4                        # learning rate
     λ = 0.01f0                      # regularization paramater
     batch_size = 64                 # minibatch size
     seq_len = 50                    # sequence length for training samples
     epochs = 800                    # number of epochs for training
-    seed = 3                        # random seed
+    seed = 1                        # random seed
     cuda = false                    # GPU usage (not working well yet)
     dt = 0.05                       # timestep for ode solve
     start_af = 0.0001f0             # Annealing factor start value
     end_af = 0.001f0                # Annealing factor end value
-    ae = 400                        # Annealing factor epoch end
+    ae = 200                        # Annealing factor epoch end
 
     ## Progressive observation training
     progressive_training = false    # progressive training usage
-    prog_training_duration = 5      # number of eppchs to reach the final seq_len
+    prog_training_duration = 200    # number of eppchs to reach the final seq_len
     start_seq_len = 10              # training sequence length at first step
 
     ## Visualization
-    vis_len = 60                    # number of frames to visualize after each epoch
+    vis_len = 60                    # number of test frames to visualize after each epoch
     save_figure = false             # true: save visualization figure in save_path folder
                                     # false: display image instead of saving it    
 end
@@ -65,36 +69,44 @@ function train(; kws...)
 
     ############################################################################
     ## Prepare training data
-
     root_dir = @__DIR__
-    data_path = "$root_dir/data/processed_data.jld2"
+    data_path = "$root_dir/data/data.bson"
 
     if ~isfile(data_path)
-        @info "Downloading pendulum data"
+        @info "Generating data"
+        latent_data, u0s, ps, high_dim_data = generate_dataset(diffeq = diffeq)
+        data = (latent_data, u0s, ps, high_dim_data)
         mkpath("$root_dir/data")
-        download("https://ndownloader.figshare.com/files/27986997", data_path)
+        @save data_path data
     end
+    
+    data_loaded = load(data_path, :data)
+    train_data = data_loaded[4]
+    latent_data = data_loaded[1]
 
-    data_loaded = load(data_path, "processed_data")
-    train_data = data_loaded["train"]
+    # Stack time for each sample
+    train_data = Flux.stack.(train_data, 3)
 
-    train_data_norm, min_val, max_val = normalize_to_unit_segment(train_data)
-    observations, full_seq_len, h, w = size(train_data_norm)
+    # Stack all samples
+    train_data = Flux.stack(train_data, 4) # 28x28x400x450
+    h, w, full_seq_len, observations = size(train_data)
+    latent_data = Flux.stack(latent_data, 3)
 
-    train_data = reshape(train_data_norm, observations, full_seq_len, :)
-    train_data = permutedims(train_data, [3, 2, 1]) # input_dim, time_size, observations
+    # Vectorize frames
+    train_data = reshape(train_data, :, full_seq_len, observations) # input_dim, time_size, samples
     train_data = Float32.(train_data)
 
-    train_set, val_set = splitobs(train_data, 0.9)
+    train_set, val_set = Array.(splitobs(train_data, 0.9))
+    train_set_latent, val_set_latent = Array.(splitobs(latent_data, 0.9))
 
-    loader_train = DataLoader(Array(train_set), batchsize=batch_size, shuffle=true, partial=false)
-    loader_val = DataLoader(Array(val_set), batchsize=size(val_set, 3), shuffle=false, partial=false)
+    # loader_train = DataLoader(train_set, batchsize=batch_size, shuffle=true, partial=false)
+    loader_train = DataLoader((train_set, train_set_latent), batchsize=batch_size, shuffle=true, partial=false)
+    val_set_time_unstacked = Flux.unstack(val_set, 2)
 
     input_dim = size(train_set,1)
 
     ############################################################################
     # Create model
-
     encoder_layers, decoder_layers = default_layers(model_type, input_dim, diffeq, device)
     model = LatentDiffEqModel(model_type, encoder_layers, diffeq, decoder_layers)
 
@@ -104,10 +116,10 @@ function train(; kws...)
     ############################################################################
     ## Define optimizer
     opt = AdaBelief(η)
+    # opt = ADAM(η)
 
     ############################################################################
     ## Various definitions
-
     if progressive_training
         prog_seq_lengths = range(start_seq_len, seq_len, step=(seq_len-start_seq_len)/(prog_training_duration-1))
         prog_seq_lengths = Int.(round.(prog_seq_lengths))
@@ -115,8 +127,10 @@ function train(; kws...)
         prog_training_duration = 0
     end
     
-    best_val_loss::Float32 = Inf32
-    val_loss::Float32 = 0
+    # Preparation for saving best models weights
+    mkpath("$root_dir/output")
+    best_val_loss = Inf32
+    val_loss = 0f0
 
     # mkpath("$root_dir/output")
     # args = struct2dict(args)
@@ -127,12 +141,13 @@ function train(; kws...)
         mkpath("$root_dir/output/visualization")
         GR.inline("pdf")
     end
+
     ############################################################################
     ## Main train loop
     @info "Start Training of $(typeof(model_type))-net, total $epochs epochs"
     for epoch = 1:epochs
 
-        ## set a sequence length for training samples
+        # Set a sequence length for training samples
         seq_len = epoch ≤ prog_training_duration ? prog_seq_lengths[epoch] : seq_len
 
         # Model evaluation length
@@ -141,8 +156,9 @@ function train(; kws...)
         mb_id = 1   # Minibatch id
         @info "Epoch $epoch .. (Sequence training length $seq_len)"
         progress = Progress(length(loader_train))
-        for x in loader_train
 
+        for data in loader_train
+            x, latent = data
             # Comput annealing factor
             af = annealing_factor(start_af, end_af, ae, epoch, mb_id, length(loader_train))
             mb_id += 1
@@ -158,24 +174,23 @@ function train(; kws...)
             Flux.Optimise.update!(opt, ps, grad)
 
             # Use validation set to get loss and visualisation
-            val_set = Flux.unstack(first(loader_val), 2)
-            t_val = range(0.f0, step=dt, length=length(val_set))
-            val_loss = loss_batch(model, λ, val_set |> device, t_val, af)
+            t_val = range(0.f0, step=dt, length=length(val_set_time_unstacked))
+            val_loss = loss_batch(model, λ, val_set_time_unstacked |> device, t_val, af)
 
-            # progress meter
+            # Progress meter
             next!(progress; showvalues=[(:loss, loss),(:val_loss, val_loss)])
         end
 
         if device != gpu
-            val_set = first(loader_val)
-            # visualize_val_image(model, val_set[:,1:vis_len,:] |> device, t_val, h, w, save_figure)
-            visualize_val_image(model, val_set |> device, vis_len, dt, h, w, save_figure)
+            visualize_val_image(model, val_set |> device, val_set_latent, vis_len, dt, h, w, save_figure)
         end
-        # if val_loss < best_val_loss
-            # best_val_loss = deepcopy(val_loss)
-            # @save "$root_dir/output/best_model.bson" model
-            # @info "Model saved"
-        # end
+
+        if (val_loss < best_val_loss) & (epoch ≥ ae)
+            best_val_loss = deepcopy(val_loss)
+            weights = Flux.params(model)
+            @save "$root_dir/output/best_model_weights.bson" weights
+            @info "Model saved"
+        end
     end
 end
 
@@ -198,15 +213,15 @@ function loss_batch(model, λ, x, t, af)
     return reconstruction_loss + af*kl_loss
 end
 
-
 ################################################################################
 ## Visualization function
 
-function visualize_val_image(model, val_set, vis_len, dt, h, w, save_figure)
+function visualize_val_image(model, val_set, val_set_latent, vis_len, dt, h, w, save_figure)
     j = rand(1:size(val_set,3))
     idxs = rand_time(size(val_set,2), vis_len)
     X_test = val_set[:, idxs, j]
-    
+    true_latent = val_set_latent[:,idxs,j]
+
     frames_test = [Gray.(reshape(x,h,w)) for x in eachcol(X_test)]
     X_test = reshape(X_test, Val(3))
     x = Flux.unstack(X_test, 2)
@@ -215,18 +230,14 @@ function visualize_val_image(model, val_set, vis_len, dt, h, w, save_figure)
     X̂, μ, logσ² = model(x, t_val)
     x̂, ẑ, ẑ₀, = X̂
 
-    if length(X̂) == 4
-        θ̂ = X̂[4]
-        @show θ̂
-    end
+    θ̂ = X̂[4]
+    println("Inferred Pendulum Length = $θ̂")
 
-    # gr(size = (700, 350))
     ẑ = Flux.stack(ẑ, 2)
-
-    plt1 = plot(ẑ[1,:,1], legend = false)
+    plt1 = plot(ẑ[1,:,1], label="inferred",legend=:topleft)
     ylabel!("Angle")
     xlabel!("time")
-    # plt1 = plot(ẑ[1,1,:]) # for Latent ODE
+    plt1 = plot!(twinx(), true_latent[1,:], color=:red, box = :on, xticks=:none, label="ground truth")
 
     x̂ = Flux.stack(x̂, 2)
     frames_pred = [Gray.(reshape(x,h,w)) for x in eachslice(x̂, dims=2)]
