+++ examples/pendulum_friction-less/model_train_original_data.jl
@@ -1,4 +1,6 @@
-using Flux: length
+# Example of GOKU-net model on the original friction-less pendulum data
+# from the  GOKU-net paper (https://github.com/orilinial/GOKU)
+
 using LatentDiffEq
 using FileIO
 using Parameters: @with_kw
@@ -10,15 +12,10 @@ using BSON: @save
 using Flux.Data: DataLoader
 using Flux
 using OrdinaryDiffEq
-using StochasticDiffEq
 using ModelingToolkit
 using Images
 using Plots
 import GR
-using DiffEqSensitivity
-
-include("FreeFall.jl")
-include("create_data_ff.jl")
 
 ################################################################################
 ## Arguments for the train function
@@ -28,8 +25,7 @@ include("create_data_ff.jl")
     # model_type = LatentODE()
 
     ## Latent Differential Equations
-    diffeq = FreeFall()
-    # diffeq = SPendulum()
+    diffeq = Pendulum()
     # diffeq = NODE(2)
 
     ## Training params
@@ -38,16 +34,16 @@ include("create_data_ff.jl")
     batch_size = 64                 # minibatch size
     seq_len = 50                    # sequence length for training samples
     epochs = 800                    # number of epochs for training
-    seed = 1                        # random seed
+    seed = 3                        # random seed
     cuda = false                    # GPU usage (not working well yet)
     dt = 0.05                       # timestep for ode solve
-    start_af = 0.01f0             # Annealing factor start value
-    end_af = 0.1f0                # Annealing factor end value
+    start_af = 0.0001f0             # Annealing factor start value
+    end_af = 0.001f0                # Annealing factor end value
     ae = 400                        # Annealing factor epoch end
 
     ## Progressive observation training
     progressive_training = false    # progressive training usage
-    prog_training_duration = 200    # number of eppchs to reach the final seq_len
+    prog_training_duration = 5      # number of eppchs to reach the final seq_len
     start_seq_len = 10              # training sequence length at first step
 
     ## Visualization
@@ -61,39 +57,9 @@ end
 ## Training done manualy
 
 function train(; kws...)
-       ## Global model
-       model_type = GOKU()
-       # model_type = LatentODE()
-   
-       ## Latent Differential Equations
-       diffeq = FreeFall()
-       # diffeq = SPendulum()
-       # diffeq = NODE(2)
-   
-       ## Training params
-       η = 1e-5                        # learning rate
-       λ = 0.01f0                      # regularization paramater
-       batch_size = 10                 # minibatch size changed from 64
-       seq_len = 50                    # sequence length for training samples
-       epochs = 800                    # number of epochs for training
-       seed = 3                        # random seed
-       cuda = false                    # GPU usage (not working well yet)
-       dt = 0.05                       # timestep for ode solve
-       start_af = 0.0001f0             # Annealing factor start value
-       end_af = 0.001f0                # Annealing factor end value
-       ae = 400                        # Annealing factor epoch end
-   
-       ## Progressive observation training
-       progressive_training = false    # progressive training usage
-       prog_training_duration = 200    # number of eppchs to reach the final seq_len
-       start_seq_len = 10              # training sequence length at first step
-   
-       ## Visualization
-       vis_len = 60                    # number of frames to visualize after each epoch
-       save_figure = false             # true: save visualization figure in save_path folder
-                                       # false: display image instead of saving it    
-   
     ## Load hyperparameters and GPU config
+    args = Args(; kws...)
+    @unpack_Args args
 
     seed > 0 && Random.seed!(seed)
 
@@ -104,40 +70,28 @@ function train(; kws...)
     ## Prepare training data
 
     root_dir = @__DIR__
-    data_path = "$root_dir/data/data.bson"
+    data_path = "$root_dir/data/processed_data.jld2"
 
-    # if ~isfile(data_path)
-    if true
-        @info "Generating data"
-        latent_data, u0s, ps, high_dim_data = generate_dataset(diffeq = diffeq)
-
-        data = (latent_data, u0s, ps, high_dim_data)
+    if ~isfile(data_path)
+        @info "Downloading pendulum data"
         mkpath("$root_dir/data")
-        @save data_path data
+        download("https://ndownloader.figshare.com/files/27986997", data_path)
     end
-    
-    data_loaded = load(data_path, :data)
-    train_data = data_loaded[4]
-    latent_data = data_loaded[1]
 
-    # stack time for each sample
-    train_data = Flux.stack.(train_data, 3)
+    data_loaded = load(data_path, "processed_data")
+    train_data = data_loaded["train"]
 
-    # stack all samples
-    train_data = Flux.stack(train_data, 4) # 28x28x400x450
-    h, w, full_seq_len, observations = size(train_data)
-    latent_data = Flux.stack(latent_data, 3)
+    train_data_norm, min_val, max_val = normalize_to_unit_segment(train_data)
+    observations, full_seq_len, h, w = size(train_data_norm)
 
-    # vectorize frames
-    train_data = reshape(train_data, :, full_seq_len, observations) # input_dim, time_size, samples
+    train_data = reshape(train_data_norm, observations, full_seq_len, :)
+    train_data = permutedims(train_data, [3, 2, 1]) # input_dim, time_size, observations
     train_data = Float32.(train_data)
 
-    train_set, val_set = Array.(splitobs(train_data, 0.9))
-    train_set_latent, val_set_latent = Array.(splitobs(latent_data, 0.9))
+    train_set, val_set = splitobs(train_data, 0.9)
 
-    # loader_train = DataLoader(Array(train_set), batchsize=batch_size, shuffle=true, partial=false)
-    loader_train = DataLoader((train_set, train_set_latent), batchsize=batch_size, shuffle=true, partial=false)
-    val_set_time_unstacked = Flux.unstack(val_set, 2)
+    loader_train = DataLoader(Array(train_set), batchsize=batch_size, shuffle=true, partial=false)
+    loader_val = DataLoader(Array(val_set), batchsize=size(val_set, 3), shuffle=false, partial=false)
 
     input_dim = size(train_set,1)
 
@@ -145,7 +99,6 @@ function train(; kws...)
     # Create model
 
     encoder_layers, decoder_layers = default_layers(model_type, input_dim, diffeq, device)
-
     model = LatentDiffEqModel(model_type, encoder_layers, diffeq, decoder_layers)
 
     # Get parameters
@@ -154,7 +107,6 @@ function train(; kws...)
     ############################################################################
     ## Define optimizer
     opt = AdaBelief(η)
-    # opt = ADAM(η)
 
     ############################################################################
     ## Various definitions
@@ -190,11 +142,10 @@ function train(; kws...)
         t = range(0.f0, step=dt, length=seq_len)
 
         mb_id = 1   # Minibatch id
-        @info "Epoch $epoch .. (Sequence training length $seq_len)" 
+        @info "Epoch $epoch .. (Sequence training length $seq_len)"
         progress = Progress(length(loader_train))
+        for x in loader_train
 
-        for data in loader_train
-            x, latent = data
             # Comput annealing factor
             af = annealing_factor(start_af, end_af, ae, epoch, mb_id, length(loader_train))
             mb_id += 1
@@ -210,18 +161,18 @@ function train(; kws...)
             Flux.Optimise.update!(opt, ps, grad)
 
             # Use validation set to get loss and visualisation
-            t_val = range(0.f0, step=dt, length=length(val_set_time_unstacked))
-            val_loss = loss_batch(model, λ, val_set_time_unstacked |> device, t_val, af)
+            val_set = Flux.unstack(first(loader_val), 2)
+            t_val = range(0.f0, step=dt, length=length(val_set))
+            val_loss = loss_batch(model, λ, val_set |> device, t_val, af)
 
             # progress meter
             next!(progress; showvalues=[(:loss, loss),(:val_loss, val_loss)])
-            # @show [(:loss, loss),(:val_loss, val_loss)]
         end
 
         if device != gpu
+            val_set = first(loader_val)
             # visualize_val_image(model, val_set[:,1:vis_len,:] |> device, t_val, h, w, save_figure)
-            visualize_val_image(model, val_set |> device, val_set_latent, vis_len, dt, h, w, save_figure)
-            # visualize_val_image(model, val_set[:,1:3,:] |> device, t_val, h, w, save_figure)
+            visualize_val_image(model, val_set |> device, vis_len, dt, h, w, save_figure)
         end
         # if val_loss < best_val_loss
             # best_val_loss = deepcopy(val_loss)
@@ -250,34 +201,15 @@ function loss_batch(model, λ, x, t, af)
     return reconstruction_loss + af*kl_loss
 end
 
-import LatentDiffEq.vector_kl
-
-function vector_kl(μ::T, logσ²::T) where T <: Tuple{Matrix, Matrix}
-    P = eltype(μ[1])
-    s = zero(P)
-    # go through initial conditions and parameters
-    @inbounds for i in 1:2
-        s1 = zero(P)
-        @inbounds for k in eachindex(μ[i])
-            s1 += kl(μ[i][k], logσ²[i][k])
-        end
-        # divide per batch size
-        s1 /= size(μ[i], 2)
-        s += s1
-    end
-    return s
-end
 
 ################################################################################
 ## Visualization function
 
-
-function visualize_val_image(model, val_set, val_set_latent, vis_len, dt, h, w, save_figure)
+function visualize_val_image(model, val_set, vis_len, dt, h, w, save_figure)
     j = rand(1:size(val_set,3))
     idxs = rand_time(size(val_set,2), vis_len)
     X_test = val_set[:, idxs, j]
-    true_latent = val_set_latent[:,idxs,j]
-
+    
     frames_test = [Gray.(reshape(x,h,w)) for x in eachcol(X_test)]
     X_test = reshape(X_test, Val(3))
     x = Flux.unstack(X_test, 2)
@@ -291,13 +223,13 @@ function visualize_val_image(model, val_set, val_set_latent, vis_len, dt, h, w, 
         @show θ̂
     end
 
+    # gr(size = (700, 350))
     ẑ = Flux.stack(ẑ, 2)
 
-    plt1 = plot(ẑ[1,:,1], label="inferred",legend=:topleft)
-    ylabel!("Height")
+    plt1 = plot(ẑ[1,:,1], legend = false)
+    ylabel!("Angle")
     xlabel!("time")
-    plt1 = plot!(twinx(), true_latent[1,:], color=:red, box = :on, xticks=:none, label="ground truth")
-
+    # plt1 = plot(ẑ[1,1,:]) # for Latent ODE
 
     x̂ = Flux.stack(x̂, 2)
     frames_pred = [Gray.(reshape(x,h,w)) for x in eachslice(x̂, dims=2)]
@@ -307,9 +239,7 @@ function visualize_val_image(model, val_set, val_set_latent, vis_len, dt, h, w, 
 
     plt2 = mosaicview(frames_test..., frames_pred..., nrow=2, rowmajor=true)
     plt2 = plot(plt2, leg = false, ticks = nothing, border = :none)
-    # plt3 = plot(ẑ[1,1,:]) # for Latent ODE
     plt = plot(plt1, plt2, layout = @layout([a; b]))
-    # plt = plot(plt3)
     save_figure ? savefig(plt, "output/visualization/fig.pdf") : display(plt)
 end
